Sets the CLREventsBreakpoint to inactive, thus preventing phantom debug sessions
This deactivates the `CLREventsBreakpoint` created in `CoreDebugProcess.StartCLR()`. This breakpoint breaks on CLR events such as throws, catches, assemblies loaded, threads started and stopped, etc. If the breakpoint is active, this causes a “phantom debug” session, in which the process is frozen, but the IDE does not expose line highlighting and step/continue functionality. The result _looks_ like a frozen system.